requirements.txt
flask
flask-sqlalchemy
flask-migrate
psycopg2-binary
flask-cors
flask-restx